Update quantitative.md
Additional comments:

For Product type - can a product be listed as more than one type?  For example, LOLA could be a geodetic coordinate reference frame and elevation data.  I have included such a note, but delete it if you disagree with the change.

Resolution and Coverage - consider stating how this is reported for consistency with other fields(e.g., from the image header?)

Does an abstract count as a reference?  They aren't peer-reviewed and a lot of products (cough...THEMIS) are not published in a peer reviewed journal.