Update _index.md
A relatively controlled image mosaic "floats" above the geodetic coordinate reference frame and has not been tied to the broader geospatial context.  - I *think* that relatively controlled images can also float below the geodetic.  DTMs before tied to ground can be higher or lower than the geodetic.  If this is a true statement, then the sentence should be corrected for improved accuracy.

A rigorously controlled image mosaic that has been rectified to a spherical body representation can suffer from significant topography-induced errors that cannot be removed simply by controlling the data, and we do not consider such an image or mosaic to be orthorectified without topographic information. - This statement isn't always correct.  For example, the difference between using topography for Europa or a sphere is so miniscule that it really does not matter.  A sphere is just as good.  This point of clarification is great.  I just think it needs to acknowledge that this strict definition doesn't always practically hold.